refactor: extract image compression helper
